PARMED: move slicing ResidueTemplate from test_parmed_residue to correct test file
